Fixes and LINT.
